Move error types into modules related to them.
