+++ smwe-rom/src/level/secondary_entrance.rs
@@ -1,8 +1,8 @@
 use crate::{
     disassembler::binary_block::{DataBlock, DataKind},
-    error::RomError,
     snes_utils::{addr::AddrSnes, rom::noop_error_mapper, rom_slice::SnesSlice},
     RomDisassembly,
+    RomError,
 };
 
 pub const SECONDARY_ENTRANCE_TABLE: SnesSlice = SnesSlice::new(AddrSnes(0x05F800), 512);
